all the test and downloading part done
