--- data/fmri_data/Makefile
@@ -1,5 +1,0 @@
-target:
-	sh downloader.sh
-
-validate:
-	python verify_data.py
