--- data/fmri_data/__init__.py
@@ -1,1 +1,0 @@
-
