--- data/fmri_data/tests/test_verify_data.py
@@ -1,16 +1,0 @@
-from __future__ import absolute_import, division, print_function
-
-import tempfile
-
-# from .. import ds105
-
-# def test_check_hashes():
-    
-#     with tempfile.NamedTemporaryFile() as temp:
-#         temp.write(b'Some data')
-#         temp.flush()
-#         fname = temp.name
-#         o = {fname: "5b82f8bf4df2bfb0e66ccaa7306fd024"}
-#         assert data.check_hashes(d)
-#         o = {fname: "4b82f8bf4df2bfb0e66ccaa7306fd024"}
-#         assert not data.check_hashes(d)
